[photon] Add ADVISORY references to VMWare Photon
Photon OS is a Linux based, open source, security-hardened,
enterprise grade appliance operating system that is purpose
built for Cloud and Edge applications.

Photon Advisories are published at
https://github.com/vmware/photon/wiki/Security-Advisories
There are 4 major versions of Photon

Photon 1 and 2  are out of support, so no further changes are expected:
https://blogs.vmware.com/vsphere/2022/01/photon-1-x-end-of-support-announcement.html

This is a one-time Pull Request for now, that adds Photon Advisories
to the references section. It goes down the list to find a suitable
place to add the links:

1. In the gsd.references section as an ADVISORY url
2. GSD.references section, as an additional URL
3. Create a new GSD.refererences section if GSD exists with list of URLs
4. Create a new GSD section, with list of URLs, picking up the
   description from the CVE/NVD descriptions